Blind75 - Reverse Linked List
The initial intuition would be to consider the nextnode as the previous node and call current.next as previous, but just before that, save the current.next to a variable called nextnode. And remember we keep changing the head, so assign that to something called current. It is a while loop for current that will return the prevnode